MYC-1317: Move login logic to root page for better code organization and reusability
